refactor(product): split terminal replay buffer
